% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DML.Processing.R
\name{DML.Processing}
\alias{DML.Processing}
\title{Data Manipulating Language}
\value{
A \code{list} of DML Commands:
\itemize{
\item \code{INSERT(keyValues)}
\item \code{VALUES(keyValues)}
\item \code{UPDATE(table)}
\item \code{SET(keyValues)}
\item \code{DELETE(keyValues)}
}
}
\description{
DML is a subset or SQL and used to manipulating data in SQL Database
}
