% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DML.Service.R
\name{DML.Service}
\alias{DML.Service}
\title{Data Manipulating Language}
\value{
A \code{list} of DML Services:
\itemize{
\item \code{INSERT()}
\item \code{INTO()}
\item \code{SCHEMA(schema)}
\item \code{TABLE(table)}
\item \code{VALUES()}
\item \code{KEYS(keyValues)}
\item \code{NEWVALUES(keyValues)}
}
}
\description{
DML is a subset or SQL and used to manipulating data in SQL Database
}
