% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQL.Processing.R
\name{DQL.Processing}
\alias{DQL.Processing}
\title{Data Query Language'}
\value{
A \code{list} of DQL Commands:
\itemize{
\item \code{SELECT(fields)}
\item \code{FROM(table)}
\item \code{WHERE(key, value)}
}
}
\description{
DQL is a subset or SQL and used to perform queries against tables in SQL Database
}
