% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartile.Exceptions.R
\name{Quartile.Exceptions}
\alias{Quartile.Exceptions}
\title{Quartile Exceptions}
\value{
A \code{list} of exceptions:
\itemize{
\item \code{NullException()}
\item \code{DataTypeException()}
}
}
\description{
\code{Quartile.Exceptions()} is an internal function that returns a set of exceptions used by \code{Quartile.Validation()} service.
\itemize{
\item The first exception: \code{NullException}, indicates to the user that no data have to provided for analysis.
\item The second exception: \code{DataTypeException}, indicates to the user than the data provided is not numerical data.
}
}
