% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartile.Validation.R
\name{Quartile.Validation}
\alias{Quartile.Validation}
\title{Quartile Validations}
\value{
A \code{list} of validators:
\itemize{
\item \code{DataExist()}
\item \code{DataIsNumeric()}
}
}
\description{
\code{Quartile.Validations()} is an internal function that returns a set of validators used by \code{Quartile.Service()}.
\itemize{
\item The first validator: \code{DataExist}, checks if a non-null data argument was provided.
\item The second validator: \code{DataIsNumeric}, checks if the data argument provided contains numerical data.
}
}
