% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.Functions.Service.R
\name{SQL.Functions.Service}
\alias{SQL.Functions.Service}
\title{SQL Functions'}
\value{
A \code{list} of SQL Functions:
\itemize{
\item \code{LOWER(value, alias)}
}
}
\description{
SQL contains a set of Utility Functions used to perform primitive operations
}
