% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.Orchestration.R
\name{SQL.Orchestration}
\alias{SQL.Orchestration}
\title{SQL Orchestration Service}
\value{
A \code{list} of services:
\itemize{
\item \code{SELECT(fields)}
\item \code{FROM(table)}
\item \code{WHERE(key, value)}
\item \code{INSERT(table, keyValues)}
\item \code{VALUES(keyValues)}
\item \code{UPDATE(table)}
\item \code{SET(keyValues)}
\item \code{DELETE()}
}
}
\description{
a collection of services used to construct SQL Queries during runtime
}
