% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.Utility.Service.R
\name{SQL.Utility.Service}
\alias{SQL.Utility.Service}
\title{Helper Utilities}
\value{
A \code{list} of Helper Utilities:
\itemize{
\item \code{BRACKET(value)}
}
}
\description{
Utility services not included in the standard set of SQL functions and used to perform primitive operations
}
