% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQL.Service.R
\name{DQL.Service}
\alias{DQL.Service}
\title{Data Query Language'}
\value{
A \code{list} of DQL Services:
\itemize{
\item \code{SELECT()}
\item \code{FROM()}
\item \code{WHERE()}
}
}
\description{
DQL is a subset or SQL and used to perform queries against tables in SQL Database
}
